added tests for day 18 answers

+++ src/aoc-2020-csharp/Day18/Day18.cs
@@ -4,7 +4,7 @@ public static class Day18
 {
     private static readonly string Input = File.ReadAllText("Day18/day18.txt");
 
-    public static int Part1() => 1;
+    public static long Part1() => 1;
 
-    public static int Part2() => 2;
+    public static long Part2() => 2;
 }
